renamed task_manager.rs to task_pool.rs
Updated readme
